generate: reintroduce looprpc
A previous PR fixed an error for `swapserverrpc` docs generation but broke `looprpc` docs generation. As a result, there are no docs for `looprpc.*` RPC calls.

The core of the issue is that if you do not specify an explicit source directory, the root of the `loop` repo is used which breaks the relative imports.

This patch adds a `sed` patcher that does an in-place replacement of these broken imports. Thus, the source directory can be `loop` itself -- readding the removed `looprpc` documentation.